requirements.txt
# Core
numpy
pandas
matplotlib
seaborn

# Image processing
opencv-python
Pillow

# Machine Learning
scikit-learn

# Deep Learning
torch
torchvision
torchaudio

# App
streamlit

# Utilities
tqdm
